Move funcs to utils.py
