+++ checkov/terraform/image_referencer/base_provider.py
@@ -8,7 +8,7 @@ from hcl2 import START_LINE, END_LINE
 from checkov.common.graph.graph_builder import CustomAttributes
 from checkov.common.images.image_referencer import Image
 from checkov.common.util.str_utils import removeprefix
-from checkov.terraform.graph_builder.graph_components.block_types import BlockType
+from checkov.terraform.graph_builder.utils import setup_file_path_to_referred_id, get_related_resource_id
 
 if TYPE_CHECKING:
     from networkx import DiGraph
@@ -46,7 +46,7 @@ class BaseTerraformProvider:
             if extract_images_func:
                 image_names.extend(extract_images_func(resource))
 
-            related_resource_id = self.get_related_resource_id(resource, file_path_to_referred_id)
+            related_resource_id = get_related_resource_id(resource, file_path_to_referred_id)
             for name in image_names:
                 images.append(
                     Image(
@@ -59,23 +59,3 @@ class BaseTerraformProvider:
                 )
 
         return images
-
-    @staticmethod
-    def get_related_resource_id(resource: dict[str, Any], file_path_to_referred_id: dict[str, str]) -> str:
-        resource_id = resource.get(CustomAttributes.ID)
-        # for external modules resources the id should start with the prefix module.[module_name]
-        if resource.get(CustomAttributes.MODULE_DEPENDENCY):
-            referred_id = file_path_to_referred_id.get(f'{resource.get(CustomAttributes.FILE_PATH)}[{resource.get(CustomAttributes.MODULE_DEPENDENCY)}#{resource.get(CustomAttributes.MODULE_DEPENDENCY_NUM)}]')
-            resource_id = f'{referred_id}.{resource_id}'
-        return resource_id
-
-
-def setup_file_path_to_referred_id(graph_object: DiGraph) -> dict[str, str]:
-    file_path_to_module_id = {}
-    modules = [node for node in graph_object.nodes.values() if
-               node.get(CustomAttributes.BLOCK_TYPE) == BlockType.MODULE]
-    for modules_data in modules:
-        for module_name, module_content in modules_data.get(CustomAttributes.CONFIG, {}).items():
-            for path in module_content.get("__resolved__", []):
-                file_path_to_module_id[path] = f"module.{module_name}"
-    return file_path_to_module_id
